Refactor out court_breach_no_payment
